#3 simple authorizations

+++ app/Http/Controllers/IdeaController.php
@@ -9,7 +9,9 @@ class IdeaController extends Controller
 {
     public function store()
     {
-        $validated = request()->validate([ 'content' => 'required|min:5|max:255', ]);
+        $validated = request()->validate(['content' => 'required|min:5|max:255',]);
+
+        $validated['user_id'] = auth()->id();
 
         Idea::create($validated);
 
@@ -23,6 +25,10 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+        if (auth()->id() !== $idea->user_id) { // If the authenticated user is not the owner of the idea
+            abort(403);
+        }
+
         $editing = true;
         return view('ideas.show',compact('idea','editing'));
     }
@@ -39,6 +45,11 @@ class IdeaController extends Controller
 
     public function destroy(Idea $idea)
     {
+        if (auth()->id() !== $idea->user_id) { // If the authenticated user is not the owner of the idea
+            abort(403);
+//            return back()->with('error', 'You are not allowed to delete this idea!');
+        }
+
         $idea->delete();
         return redirect()->route('dashboard')->with('success', 'Idea deleted successfully!');
     }
